Remove .env file from .gitignore
Having discussed it with Adobe's Gabriel Walt, I'm removing the .env
file from the .gitignore. It was originally placed here in the template
sources for historical reasons. The .env files used to contain secrets,
which can now be managed via GitHub. There's no longer a good reason not
to share this file via the Git repository.